+++ yt/yt/server/node/exec_node/layer_location.h
@@ -58,8 +58,8 @@ public:
         NProfiling::TEventTimerGuard volumeCreateTimeGuard,
         int userId,
         const std::optional<std::string>& placePath,
-        std::optional<int> diskSpaceLimit,
-        std::optional<int> inodeLimit,
+        std::optional<i64> diskSpaceLimit,
+        std::optional<i64> inodeLimit,
         const std::vector<TOverlayData>& overlayDataArray,
         bool placeInUserSlot);
 
@@ -236,8 +236,8 @@ private:
         NProfiling::TEventTimerGuard volumeCreateTimeGuard,
         int userId,
         const std::optional<std::string>& placePath,
-        std::optional<int> diskSpaceLimit,
-        std::optional<int> inodeLimit,
+        std::optional<i64> diskSpaceLimit,
+        std::optional<i64> inodeLimit,
         const std::vector<TOverlayData>& overlayDataArray,
         bool placeInUserSlot);
